drop action_id

+++ models/silver/transfers/non_native/silver__token_transfer_mints.sql
@@ -68,7 +68,7 @@ ft_mints_final AS (
             input => log_data :data
         ) f
     WHERE
-        amount_unadj :: INT > 0
+        amount_unadj :: DOUBLE > 0
 )
 SELECT
     block_timestamp,
